db: udf capability gating, dialect-doc decisions, structural cleanups (§2.5-2.7)
§2.5 - gate the udf sub-object on a backend capability. New
'unsigned char supports_udf' on the vtable (SQLite 1, Postgres 0); the
connection object attaches 'udf' only when the backend supports it, so a
Postgres connection has no udf sub-object at all (presence is the capability
probe: conn.udf ~= nil / !!conn.udf) instead of a method that fails at call
time. Verified on a real PG backend in e2e_postgres.sh (pg_has_udf=false).

§2.6 - autoincrement_id_ddl stays app-facing, now documented as an escape
hatch (CLAUDE.md): the recommended path for portable schema is migrations, not
interpolating the dialect DDL fragment. Removing it would break the stdlib's
portable CREATE TABLE (audit-log, outbox) for no real gain.

§2.7 - structural cleanups:
  - hl_cap_db_check_namespace (the backend-agnostic _hull_* guard) moved out of
    the DSN selector (db_select.c) to a new cap/db_common.c.
  - the $VAR env-ref parser moved from manifest.h to a neutral
    include/hull/utils/env_ref.h (hl_manifest_env_ref -> hl_env_ref), so the
    domain-free utils/host_match.c no longer includes manifest.h for it (the
    coupling flagged in §2.8). Callers (db_registry.c, host_match.c, tests)
    updated.
  - deferred: folding cap/db.c (SQLite engine) into cap/db_sqlite.c (adapter).
    A ~400-line mechanical move with organizational-only value; the
    engine/adapter split is defensible layering. Best done when onboarding the
    next backend (statics are already clash-free, so it stays trivial).

Full unit suite + all flavor builds (default, DB=0, postgres-only,
pure-compute) pass; postgres e2e green incl. the new udf-gate assertion.
