+++ src/hull/cap/db_common.c
@@ -1,0 +1,34 @@
+/*
+ * cap/db_common.c - backend-agnostic database helpers
+ *
+ * Shared db helpers with no backend or SQL-engine dependency, so they compile
+ * in every DB flavor (including Postgres-only). Currently the _hull_* namespace
+ * guard, previously wedged into the DSN selector (db_select.c); a natural home
+ * for future dialect-neutral helpers.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+#ifdef HL_ENABLE_DB
+
+#include "hull/cap/db.h"
+#include <strings.h>
+
+/* ── Namespace protection ──────────────────────────────────────────── */
+
+/* Backend-agnostic string check (no SQL execution): reject any SQL touching a
+ * reserved `_hull_*` table. User code is gated by a call-stack check at each
+ * binding site; stdlib bypasses via its `hull.` / `hull:` source prefix. */
+int hl_cap_db_check_namespace(const char *sql)
+{
+    if (!sql)
+        return HL_DB_ERR_DENIED;
+    for (const char *p = sql; *p; p++) {
+        if ((*p == '_' || *p == 'H' || *p == 'h') &&
+            strncasecmp(p, "_hull_", 6) == 0)
+            return HL_DB_ERR_DENIED;
+    }
+    return HL_DB_OK;
+}
+
+#endif /* HL_ENABLE_DB */
