+++ src/hull/cap/db_select.c
@@ -108,21 +108,7 @@ const HlDbBackend *hl_db_backend_select(const char *dsn, const char **err)
 #endif
 }
 
-/* ── Namespace protection ──────────────────────────────────────────── */
-
-/* Backend-agnostic string check (no SQL execution), so it lives with the
- * selector rather than the SQLite engine (cap/db.c) and stays available in a
- * Postgres-only build. Rejects any SQL touching a reserved `_hull_*` table. */
-int hl_cap_db_check_namespace(const char *sql)
-{
-    if (!sql)
-        return HL_DB_ERR_DENIED;
-    for (const char *p = sql; *p; p++) {
-        if ((*p == '_' || *p == 'H' || *p == 'h') &&
-            strncasecmp(p, "_hull_", 6) == 0)
-            return HL_DB_ERR_DENIED;
-    }
-    return HL_DB_OK;
-}
+/* hl_cap_db_check_namespace (the backend-agnostic _hull_* guard) moved to
+ * cap/db_common.c (§2.7); it was never part of DSN selection. */
 
 #endif /* HL_ENABLE_DB */
